Added functionality for painting on the window
The main concept i used to complete the drawing via mouse was the use of 2 main windows. By creating a 2nd class for the canvas, I am able to open another new window and have it's own separate dropdowns for recognising the model, saving files, and clearing the canvas.

Additionally, I also added some new icons and tidied up the code with comments.

Actions were added to the 'Save' and 'Clear'. Linear Model and recognition have yet to be mapped, but that will be done soon as the code has already been completed.