Fixed viewing test/train images
The view of the train and test data has now been fixed. It can now show all 60,000 loaded values within the MNIST dataset for both test and train folders.

The implementation was a bit intense in terms of code, but was required due to using a 'page' style instead of a scroll bar. In my opinion, this allowed the user to see many more images for each sector using next and previous buttons.