Major GUI change for drawing canvas
Instead of opening a new window, the canvas and options can now be accessed directly from the main window.

Additionally, the predicted number that was drawn will display in a read-only fashion.

An entire class for the drawing canvas was removed due to being integrated within the main window. This saved 100's of lines of code.

Lastly, a button was added to show the entire probability graph for the digit drawn if the user desires. This opens in a new window.